рефакторинг, исправляет по замечаниям пр12
