+++ src/api.js
@@ -1,14 +1,8 @@
 import TaskModel from "./models/task";
-
-const Method = {
-  GET: `GET`,
-  POST: `POST`,
-  PUT: `PUT`,
-  DELETE: `DELETE`
-};
+import {CodesResponse, Method} from "./utils/const";
 
 const checkStatus = (response) => {
-  if (response.status >= 200 && response.status < 300) {
+  if (response.status >= CodesResponse.OK && response.status < CodesResponse.MULTIPLE_CHOICE) {
     return response;
   } else {
     throw new Error(`${response.status}: ${response.statusText}`);
